requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
orjson==3.10.7


